requirements.txt
fastapi
uvicorn[standard]
google-cloud-firestore
google-cloud-aiplatform >= 1.38.0 # Ensure version supports Gemini
firebase-admin
pydantic
python-dotenv
google-auth # Often needed implicitly by google cloud libs